+++ src/Utility/IndexedArray.h
@@ -5,7 +5,7 @@
 #include <type_traits>
 #include <utility>
 
-#include "Workaround.h"
+#include "Utility/Workaround/ToUnderlying.h"
 
 
 namespace detail {
